[NUI] Make AccessibilityManager APIs as public ones (#1482)
- Move AccessibilityManager class from internal to public
- Make Tizen.NUI.Accessibility namespace
