+++ src/Tizen.NUI/src/internal/AccessibilityFocusOvershotSignal.cs
@@ -18,6 +18,7 @@
 namespace Tizen.NUI
 {
     using Tizen.NUI.BaseComponents;
+    using Tizen.NUI.Accessibility;
     internal class AccessibilityFocusOvershotSignal : Disposable
     {
 
@@ -28,19 +29,19 @@ namespace Tizen.NUI
 
         protected override void ReleaseSwigCPtr(System.Runtime.InteropServices.HandleRef swigCPtr)
         {
-            Interop.AccessibilityManage.delete_AccessibilityFocusOvershotSignal(swigCPtr);
+            Interop.AccessibilityManager.delete_AccessibilityFocusOvershotSignal(swigCPtr);
         }
 
         public bool Empty()
         {
-            bool ret = Interop.AccessibilityManage.AccessibilityFocusOvershotSignal_Empty(swigCPtr);
+            bool ret = Interop.AccessibilityManager.AccessibilityFocusOvershotSignal_Empty(swigCPtr);
             if (NDalicPINVOKE.SWIGPendingException.Pending) throw NDalicPINVOKE.SWIGPendingException.Retrieve();
             return ret;
         }
 
         public uint GetConnectionCount()
         {
-            uint ret = Interop.AccessibilityManage.AccessibilityFocusOvershotSignal_GetConnectionCount(swigCPtr);
+            uint ret = Interop.AccessibilityManager.AccessibilityFocusOvershotSignal_GetConnectionCount(swigCPtr);
             if (NDalicPINVOKE.SWIGPendingException.Pending) throw NDalicPINVOKE.SWIGPendingException.Retrieve();
             return ret;
         }
@@ -49,7 +50,7 @@ namespace Tizen.NUI
         {
             System.IntPtr ip = System.Runtime.InteropServices.Marshal.GetFunctionPointerForDelegate<System.Delegate>(func);
             {
-                Interop.AccessibilityManage.AccessibilityFocusOvershotSignal_Connect(swigCPtr, new System.Runtime.InteropServices.HandleRef(this, ip));
+                Interop.AccessibilityManager.AccessibilityFocusOvershotSignal_Connect(swigCPtr, new System.Runtime.InteropServices.HandleRef(this, ip));
                 if (NDalicPINVOKE.SWIGPendingException.Pending) throw NDalicPINVOKE.SWIGPendingException.Retrieve();
             }
         }
@@ -58,18 +59,18 @@ namespace Tizen.NUI
         {
             System.IntPtr ip = System.Runtime.InteropServices.Marshal.GetFunctionPointerForDelegate<System.Delegate>(func);
             {
-                Interop.AccessibilityManage.AccessibilityFocusOvershotSignal_Disconnect(swigCPtr, new System.Runtime.InteropServices.HandleRef(this, ip));
+                Interop.AccessibilityManager.AccessibilityFocusOvershotSignal_Disconnect(swigCPtr, new System.Runtime.InteropServices.HandleRef(this, ip));
                 if (NDalicPINVOKE.SWIGPendingException.Pending) throw NDalicPINVOKE.SWIGPendingException.Retrieve();
             }
         }
 
         public void Emit(View arg1, AccessibilityManager.FocusOvershotDirection arg2)
         {
-            Interop.AccessibilityManage.AccessibilityFocusOvershotSignal_Emit(swigCPtr, View.getCPtr(arg1), (int)arg2);
+            Interop.AccessibilityManager.AccessibilityFocusOvershotSignal_Emit(swigCPtr, View.getCPtr(arg1), (int)arg2);
             if (NDalicPINVOKE.SWIGPendingException.Pending) throw NDalicPINVOKE.SWIGPendingException.Retrieve();
         }
 
-        public AccessibilityFocusOvershotSignal() : this(Interop.AccessibilityManage.new_AccessibilityFocusOvershotSignal(), true)
+        public AccessibilityFocusOvershotSignal() : this(Interop.AccessibilityManager.new_AccessibilityFocusOvershotSignal(), true)
         {
             if (NDalicPINVOKE.SWIGPendingException.Pending) throw NDalicPINVOKE.SWIGPendingException.Retrieve();
         }
